Fixed qls grammar; introduced a saveFile dialog; changed slider and spinbox to have decimal min max and step values

+++ bg-nv/qls/src/qls/ast/rule/widget/Slider.java
@@ -2,12 +2,14 @@ package qls.ast.rule.widget;
 
 import ql.ast.type.Type;
 
+import java.math.BigDecimal;
+
 /**
  * Created by bore on 08/03/15.
  */
 public class Slider extends NumWidgetValue
 {
-    public Slider(Integer min, Integer max, Integer step)
+    public Slider(BigDecimal min, BigDecimal max, BigDecimal step)
     {
         super("slider", min, max, step);
     }
